+++ demo/models.py
@@ -1,4 +1,5 @@
 from django.db import models
+from .common import COLORS
 
 
 class Person(models.Model):
@@ -8,4 +9,13 @@ class Person(models.Model):
 
     def __unicode__(self):
         return " ".join((self.first_name, self.last_name))
+    __str__ = __unicode__
+
+
+class FavoriteColor(models.Model):
+    person = models.ForeignKey(Person)
+    color = models.CharField(max_length=100, choices=COLORS)
+
+    def __unicode__(self):
+        return "{}'s favorite color is {}".format(self.person, self.color)
     __str__ = __unicode__